Added roadmap to todo
Added a "roadmap" section to the todo.txt file.
Concept: give some idea of where the project is heading, but not so
close as to be worthy of the to-do list.
These would be generally big topic areas that have been more or less
committed to doing, but the specifics haven't been figured out.

Also included an "Ideas" section within the roadmap.
Concept: as new directions are thought of, add them here.  They can be
vague, grand, whatever.  Just write them down.